correzioni test fstcut 4.1.6

+++ FomMonitoringCore/Framework/Model/ParameterMachineModel.cs
@@ -52,7 +52,7 @@ namespace FomMonitoringCore.Framework.Model
                 enUnitaMisuraTime hmi = enUnitaMisuraTime.S;
                 if (Enum.TryParse(CnUm, true, out cn) && Enum.TryParse(HmiUm, true, out hmi))
                 {
-                    d = (temp * ((int)hmi / (int)cn)).RoundToInt();
+                    d = (temp * ((int)hmi / (double)cn)).RoundToInt();
                 }
                 TimeSpan timeSpan;
                 switch (hmi)
